+++ src/lib/data/achievements.ts
@@ -251,8 +251,8 @@ function createElectronizesAchievements(): Achievement[] {
 	return [1, 2, 3, 5, 10, 20, 50, 100, 250, 500, 1000].map(createElectronizesAchievement);
 }
 
-function createBonusPhotonClicksAchievements(): Achievement[] {
-	function createBonusPhotonClicksAchievement(count: number): Achievement {
+function createBonusHiggsBosonClicksAchievements(): Achievement[] {
+	function createBonusHiggsBosonClicksAchievement(count: number): Achievement {
 		const countNames: Record<number, string> = {
 			1: 'First',
 			10: 'Ten',
@@ -262,15 +262,15 @@ function createBonusPhotonClicksAchievements(): Achievement[] {
 		};
 
 		return {
-			id: `bonus_photons_clicked_${count}`,
-			name: `${countNames[count]} Bonus Photon`,
-			description: `Click ${formatNumber(count, 0)} bonus photon${count === 1 ? '' : 's'}`,
-			condition: (manager: GameManager) => manager.totalBonusPhotonsClicked >= count,
-			hiddenCondition: (manager: GameManager) => manager.totalBonusPhotonsClicked === 0,
+			id: `bonus_higgs_boson_clicked_${count}`,
+			name: `${countNames[count]} Bonus Higgs Boson`,
+			description: `Click ${formatNumber(count, 0)} bonus higgs boson${count === 1 ? '' : 's'}`,
+			condition: (manager: GameManager) => manager.totalBonusHiggsBosonClicked >= count,
+			hiddenCondition: (manager: GameManager) => manager.totalBonusHiggsBosonClicked === 0,
 		};
 	}
 
-	return [1, 10, 64, 512, 4096].map(createBonusPhotonClicksAchievement);
+	return [1, 10, 64, 512, 4096].map(createBonusHiggsBosonClicksAchievement);
 }
 
 function createPhotonAchievements(): Achievement[] {
@@ -317,7 +317,7 @@ const achievementsArray: Achievement[] = [
 	...createTotalLevelsAchievements(),
 	...createProtonisesAchievements(),
 	...createElectronizesAchievements(),
-	...createBonusPhotonClicksAchievements(),
+	...createBonusHiggsBosonClicksAchievements(),
 	...createPhotonAchievements(),
 	...createPhotonUpgradeAchievements(),
 	...SPECIAL_ACHIEVEMENTS,